# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Connector service, which handles any kind of external integrations for Dengankarya.",
        "title": "Connector API Docs",
        "contact": {
            "name": "Dengan Karya Support",
            "url": "https://dengankarya.com/contact",
            "email": "dukungan@dengankarya.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/payments": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new Xendit payment session (invoice) for a tenant order. Returns a checkout_url the customer visits to complete payment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment session",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tenant ID",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payment creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_payment.CreatePaymentBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment session created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_payment_domain.PaymentTransaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/shipments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a cursor-paginated list of shipments for the tenant. Supports filtering by status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "List shipments",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tenant ID",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor returned by previous response",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated statuses to filter (e.g. in_transit,delivered)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated shipment list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_common.PaginationResponse-github_com_dengankarya_connector_internal_shipping_domain_Shipment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or cursor",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new draft shipment order via Biteship.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shipping"
                ],
                "summary": "Create shipment",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Tenant ID",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shipment creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.CreateShipmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shipment created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.Shipment"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_common.PaginationResponse-github_com_dengankarya_connector_internal_shipping_domain_Shipment": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.Shipment"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_payment_domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "awaiting_payment",
                "paid",
                "settled",
                "refunding",
                "refunded",
                "expired",
                "failed",
                "voided"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusAwaitingPayment",
                "StatusPaid",
                "StatusSettled",
                "StatusRefunding",
                "StatusRefunded",
                "StatusExpired",
                "StatusFailed",
                "StatusVoided"
            ]
        },
        "github_com_dengankarya_connector_internal_payment_domain.PaymentTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "gross amount (merchant_amount + platform_fee + shipping_fee)",
                    "type": "integer"
                },
                "checkout_url": {
                    "description": "Xendit invoice payment page URL returned to the caller",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "estimated_settlement_time": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "merchant_amount": {
                    "description": "Amount - PlatformFee - ShippingFee",
                    "type": "integer"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "order_number": {
                    "description": "Tokokarya order number (string, not UUID)",
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_channel": {
                    "description": "e.g. \"BRI\", \"MANDIRI\", \"OVO\"",
                    "type": "string"
                },
                "payment_method": {
                    "description": "e.g. \"BANK_TRANSFER\", \"QRIS\", \"CREDIT_CARD\"",
                    "type": "string"
                },
                "platform_fee": {
                    "type": "integer"
                },
                "provider": {
                    "description": "\"xendit\", \"midtrans\", etc.",
                    "type": "string"
                },
                "provider_invoice_id": {
                    "description": "Xendit payment_session_id (ps-xxx)",
                    "type": "string"
                },
                "provider_payment_id": {
                    "description": "Xendit payment_id (py-xxx); populated on payment.capture",
                    "type": "string"
                },
                "settled_at": {
                    "type": "string"
                },
                "shipping_fee": {
                    "description": "shipping credit topped up to merchant's balance",
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/github_com_dengankarya_connector_internal_payment_domain.PaymentStatus"
                },
                "tenant_id": {
                    "type": "integer"
                },
                "third_party_wht": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "vat": {
                    "type": "integer"
                },
                "version": {
                    "description": "optimistic lock version; increment on every write",
                    "type": "integer"
                },
                "xendit_fee": {
                    "description": "Fee breakdown — populated by the settlement sync job after Xendit confirms settlement.",
                    "type": "integer"
                },
                "xendit_withholding_tax": {
                    "type": "integer"
                }
            }
        },
        "github_com_dengankarya_connector_internal_shipping_domain.Shipment": {
            "type": "object",
            "properties": {
                "confirmed_at": {
                    "type": "string"
                },
                "courier_code": {
                    "type": "string"
                },
                "courier_service_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_number": {
                    "description": "OrderNumber is the tokokarya's order number.",
                    "type": "string"
                },
                "picked_up_at": {
                    "type": "string"
                },
                "provider": {
                    "description": "Current default is `biteship`.",
                    "type": "string"
                },
                "provider_draft_order_id": {
                    "type": "string"
                },
                "provider_order_id": {
                    "type": "string"
                },
                "shipping_cost": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_domain.ShipmentStatus"
                },
                "tenant_id": {
                    "description": "TenantID is the merchant/tenant that owns this shipment.",
                    "type": "integer"
                },
                "tracking_number": {
                    "type": "string"
                },
                "tracking_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_shipping_provider.CourierSelection": {
            "type": "object",
            "properties": {
                "company": {
                    "description": "Company is the courier provider code.\n\nExample values:\n  - jne\n  - sicepat\n  - anteraja",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the courier service type.\n\nExample values:\n  - reg\n  - yes\n  - instant",
                    "type": "string"
                }
            }
        },
        "github_com_dengankarya_connector_internal_shipping_provider.CreateShipmentRequest": {
            "type": "object",
            "properties": {
                "courier": {
                    "description": "Courier contains the selected courier service.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.CourierSelection"
                        }
                    ]
                },
                "delivery": {
                    "description": "Delivery contains shipment delivery configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.DeliveryOptions"
                        }
                    ]
                },
                "destination": {
                    "description": "Destination contains receiver or delivery address information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.Address"
                        }
                    ]
                },
                "items": {
                    "description": "Items contains the list of shipment items.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.ShipmentItem"
                    }
                },
                "metadata": {
                    "description": "Metadata contains arbitrary internal metadata.\n\nThis field may be forwarded to providers that support metadata storage.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "note": {
                    "description": "Note contains additional shipment notes.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin contains pickup or sender address information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.Address"
                        }
                    ]
                },
                "reference_id": {
                    "description": "ReferenceID is the internal order or shipment identifier.\n\nProviders may use this value as an idempotency key or external reference.",
                    "type": "string"
                },
                "shipper": {
                    "description": "Shipper contains optional shipper or merchant information.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.Shipper"
                        }
                    ]
                },
                "tags": {
                    "description": "Tags contains optional provider-level tagging information.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_dengankarya_connector_internal_shipping_provider.DeliveryOptions": {
            "type": "object",
            "properties": {
                "cod": {
                    "description": "COD contains optional cash-on-delivery configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.CODOptions"
                        }
                    ]
                },
                "date": {
                    "description": "Date is the optional scheduled delivery date and time.\n\nThis field is primarily used for scheduled deliveries.",
                    "type": "string"
                },
                "insurance_amount": {
                    "description": "InsuranceAmount is the declared insurance value\nin the smallest currency unit.",
                    "type": "integer"
                },
                "proof_of_delivery": {
                    "description": "ProofOfDelivery contains optional proof-of-delivery configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dengankarya_connector_internal_shipping_provider.ProofOfDeliveryOptions"
                        }
                    ]
                },
                "type": {
                    "description": "Type is the shipment delivery type.\n\nExample values:\n  - now\n  - scheduled",
                    "type": "string"
                },
                "use_insurance": {
                    "description": "UseInsurance determines whether shipment insurance is enabled.",
                    "type": "boolean"
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_shipping_provider.ProofOfDeliveryOptions": {
            "type": "object",
            "properties": {
                "note": {
                    "description": "Note contains additional instructions\nfor proof-of-delivery handling.",
                    "type": "string"
                }
            }
        },
...
        "github_com_dengankarya_connector_internal_shipping_provider.Shipper": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email is the optional shipper email address.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the shipper contact name.",
                    "type": "string"
                },
                "organization": {
                    "description": "Organization is the optional organization or company name.",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone is the shipper contact phone number.",
                    "type": "string"
                }
            }
        },
...
        "internal_payment.CreatePaymentBody": {
            "type": "object",
            "properties": {
                "allowed_payment_channels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "amount": {
                    "type": "integer"
                },
                "cancel_return_url": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "customer_reference_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "order_number": {
                    "type": "string"
                },
                "platform_fee": {
                    "type": "integer"
                },
                "success_return_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    },
...
}